PR feedback: Move rpc base classes out of __init__
